Refactored a bit. Now works with the babel patch suggested in https://phabricator.babeljs.io/T7151
