+++ lib/UnexpectedMarkdown.js
@@ -1,3 +1,4 @@
+/*global JSON*/
 var extend = require('./extend');
 var Snippets = require('./Snippets');
 var createExpect = require('./createExpect');
@@ -5,12 +6,59 @@ var passError = require('passerror');
 var snippetRegexp = require('./snippetRegexp');
 var marked = require('marked-papandreou');
 var fs = require('fs');
+var pathModule = require('path');
 var convertMarkdownToMocha = require('./convertMarkdownToMocha');
+
+var fs = require('fs');
+var babelCore;
+try {
+    babelCore = require.main.require('babel-core');
+} catch (e) {}
+
+var babelOptions;
+if (babelCore) {
+    var dirFragments = pathModule.dirname(require.main.filename).split(pathModule.sep).slice(1);
+    for (var i = dirFragments.length ; i >= 0 ; i -= 1) {
+        var tryBabelRcPath = pathModule.resolve('/' + dirFragments.slice(0, i).join(pathModule.sep), '.babelrc');
+        var stats = null;
+        try {
+            stats = fs.statSync(tryBabelRcPath);
+        } catch (e) {}
+        if (stats && stats.isFile()) {
+            babelOptions = JSON.parse(fs.readFileSync(tryBabelRcPath, 'utf-8'));
+            break;
+        }
+    }
+}
+
+var maps = {};
 
 // Make the md-to-mocha transpiler avaliable as a side effect of requiring unexpected-markdown
 // so that mocha --compilers md:unexpected-markdown will work:
 require.extensions['.md'] = function (module, fileName) {
-    module._compile(convertMarkdownToMocha(fs.readFileSync(fileName, 'utf-8'), fileName), fileName);
+    var sourceMapWithCode = convertMarkdownToMocha(fs.readFileSync(fileName, 'utf-8'), fileName);
+
+    var absoluteFileName = pathModule.resolve(process.cwd(), fileName);
+
+    // Register the source map with the main UnexpectedMarkdown, which tries to detect the source-map-support
+    // module and registers a handler with it:
+    var map = sourceMapWithCode.map.toString();
+    var code = sourceMapWithCode.code;
+
+    if (babelCore) {
+        var inputSourceMap = JSON.parse(map);
+        inputSourceMap.file = fileName;
+        var babelResult = babelCore.transform(code, extend({}, babelOptions, {
+            sourceMaps: true,
+            inputSourceMap: inputSourceMap
+        }));
+        code = babelResult.code;
+        map = babelResult.map;
+    }
+
+    maps[absoluteFileName] = map;
+
+    module._compile(code, fileName);
 };
 
 var sourceMapSupport;
@@ -24,8 +72,7 @@ if (sourceMapSupport) {
     sourceMapSupport.install({
         handleUncaughtExceptions: false,
         retrieveSourceMap: function (source) {
-            var map = UnexpectedMarkdown.maps[source];
-            return map ? { url: null, map: map } : null;
+            return maps[source] ? { url: null, map: maps[source] } : null;
         }
     });
 }
@@ -38,8 +85,6 @@ function UnexpectedMarkdown(content) {
     }
     this.content = content;
 }
-
-UnexpectedMarkdown.maps = {};
 
 UnexpectedMarkdown.prototype.toHtml = function (options, cb) {
     var content = this.content;
