+++ lib/convertMarkdownToMocha.js
@@ -1,7 +1,7 @@
-/*global unexpected:true, JSON, __dirname*/
+/*global unexpected:true*/
+var pathModule = require('path');
 var esprima = require('esprima');
 var escodegen = require('escodegen');
-var pathModule = require('path');
 var snippetRegexp = require('./snippetRegexp');
 
 function extend(target) {
@@ -12,28 +12,6 @@ function extend(target) {
         });
     }
     return target;
-}
-
-var fs = require('fs');
-var babelCore;
-try {
-    babelCore = require.main.require('babel-core');
-} catch (e) {}
-
-var babelOptions;
-if (babelCore) {
-    var dirFragments = pathModule.dirname(require.main.filename).split(pathModule.sep).slice(1);
-    for (var i = dirFragments.length ; i >= 0 ; i -= 1) {
-        var tryBabelRcPath = pathModule.resolve('/' + dirFragments.slice(0, i).join(pathModule.sep), '.babelrc');
-        var stats = null;
-        try {
-            stats = fs.statSync(tryBabelRcPath);
-        } catch (e) {}
-        if (stats && stats.isFile()) {
-            babelOptions = JSON.parse(fs.readFileSync(tryBabelRcPath, 'utf-8'));
-            break;
-        }
-    }
 }
 
 function parseBlockInfo(lang) {
@@ -64,7 +42,7 @@ function parseFlags(flagsString) {
 }
 
 function parseFunctionBody(fn) {
-    return esprima.parse(fn.toString(), { loc: true, source: '<generated code>' }).body[0].body.body;
+    return esprima.parse(fn.toString()).body[0].body.body;
 }
 
 function instrumentReturns(astNode, exampleNumber) {
@@ -247,24 +225,5 @@ module.exports = function (mdSrc, fileName) {
             compileCodeBlocksToAst(codeBlocks.slice(0, i + 1), fileName)
         );
     });
-    var absoluteFileName = pathModule.resolve(process.cwd(), fileName);
-    var sourceMapWithCode = escodegen.generate(ast, { sourceMap: true, sourceMapWithCode: true });
-
-    // Register the source map with the main UnexpectedMarkdown, which tries to detect the source-map-support
-    // module and registers a handler with it:
-    var map = sourceMapWithCode.map.toString();
-    var code = sourceMapWithCode.code;
-
-    if (babelCore) {
-        var babelResult = babelCore.transform(code, extend({}, babelOptions, {
-            sourceMaps: true,
-            inputSourceMap: JSON.parse(map)
-        }));
-        code = babelResult.code;
-        map = babelResult.map;
-    }
-
-    require('../').maps[absoluteFileName] = map;
-
-    return code;
+    return escodegen.generate(ast, { sourceMap: true, sourceMapWithCode: true });
 };
